implement tide coupling between dune and XBeach
git-svn-id: https://svn.oss.deltares.nl/repos/xbeach/trunk@2557 86136381-33bf-4961-916e-47ea43f679af

+++ libxbeach.F90
@@ -429,7 +429,7 @@ contains
     ! String
     character(kind=c_char,len=*),intent(in) :: name
     ! use inout otherwise things break
-    real(c_double), intent(inout) :: x(:,:)
+    integer(c_int), intent(inout) :: x(:,:)
 
     type(arraytype) :: array
     integer :: i, index
